FT-539 new propositions (#1246)
* FT-539 correct existing propositions in mirage

Mirage has not been kept up to date when changes have been made to the
api. Also there was a lot of logic in the factories to try and reduce
the noise in the scenario definition and test case setup this was
counter productive as it was hiding what was really going on so now we
are going for full verbosity and pretty dumb factories.

* FT-539 add mirage data for new definitions

* FT-539 madhur confirmed value format for date relative

* FT-539 add new operator translations

* FT-539 fix mis-translations

* FT-539 add support for date relative

I have changed the response format of values and asked if madhur will
update the api.

* FT-539 add support for boolean_true and boolean_false

* FT-539 add support for time

* FT-539 add support for multiple

* FT-539 add support for integer

* FT-539 add support for float

* FT-539 no client side validation for now

* FT-539 explictly declare mirage test setup in case list test

* FT-539 updated login to mirror the real api more accurately

* FT-539 added missing translation

* FT-539 fix spacing of time input label

* FT-539 allow date-relative values to be selected and translate options
 Author:The Game (develop)


Former-commit-id: 0f72628c7e473c16ca0732d2f29c3131fa55e8aa
Former-commit-id: d4c00d3420296c75b56fb307ddb22d6d4754b32a [formerly c87d74595bfd0aabd7fd33b942b209027d4e6baf]
Former-commit-id: 4522e58a88d0e9370627deb625e7c09996f34078