Merge pull request #3008 from kayako/FT-2244
FT-2244: Fix file upload system for user and org views + relevant refactors.
Going to merge this as it's a big improvement but there are still issues that existed prior to this PR about uploading a file, clearing it and then uploading it again. You get to select the file but it is never attached. Author:The Game (develop)